GPS tracking now implemented in main menu and switch in the EditEM will request permission but doesn't actually activate or deactivate GPS for that EM. The EMs will all send Coordinates in the SMS when called by the keyphrase and the menu GPS tracking is on. The SMS will send a link to both google and apple maps of location. It will also send a timestamp of when the last gps coordinates were taken.
