+++ app/src/main/java/com/metrostateics499/vre_app/utility/LocationGPS.kt
@@ -27,7 +27,6 @@ class LocationGPS(context: Context) {
     }
     fun startLocationTracking(locationCallback: LocationCallback) {
         if (!startedLocationTracking) {
-            //noinspection MissingPermission
             fusedLocationClient.requestLocationUpdates(
                 locationRequest,
                 locationCallback,
