--- app/src/main/java/com/metrostateics499/vre_app/utility/TestListenKeyPhrasePopUp.kt
@@ -1,91 +1,0 @@
-package com.metrostateics499.vre_app.utility
-
-import android.annotation.SuppressLint
-import android.app.AlertDialog
-import android.app.Dialog
-import android.graphics.Color
-import android.os.Bundle
-import android.widget.EditText
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatDialogFragment
-import com.metrostateics499.vre_app.R
-
-class TestListenKeyPhrasePopUp(
-    private val textViewSelected: String,
-    private val buttonType: String
-) : AppCompatDialogFragment() {
-
-    private var editTextKeyphrase: EditText? = null
-//    private var listener: Listener? = null
-
-    @SuppressLint("CutPasteId")
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val builder = AlertDialog.Builder(activity)
-        val inflater = requireActivity().layoutInflater
-
-        when (buttonType) {
-//            "waiting" -> {
-//                val view = inflater.inflate(R.layout.layout_text_view_popup, null)
-// //                val textView: TextView = view.findViewById(R.id.text)
-// //                textView.text = textViewSelected
-//                builder.setView(view)
-//                    .setTitle("Waiting for speech")
-//                    .setNegativeButton("stop") { dialogInterface, i -> }
-// //                    .setPositiveButton("sta") { dialogInterface, i ->
-// //                        val keyphraseString = editTextKeyphrase!!.text.toString()
-// //                        listener!!.editKeyPhrase(keyphraseString)
-// //                    }
-// //                editTextKeyphrase = view.findViewById(R.id.edit_text)
-//            }
-//            "listening" -> {
-//                val view = inflater.inflate(R.layout.layout_text_view_popup, null)
-// //                val textView: TextView = view.findViewById(R.id.text)
-// //                textView.hint = "Key Phrase"
-//                builder.setView(view)
-//                    .setTitle("New Key Phrase")
-//                    .setNegativeButton("stop") { dialogInterface, i -> }
-// //                    .setPositiveButton("ok") { dialogInterface, i ->
-// //                        val keyphraseString = editTextKeyphrase!!.text.toString()
-// //                        listener!!.addKeyPhrase(keyphraseString)
-// //                    }
-// //                editTextKeyphrase = view.findViewById(R.id.edit_text)
-//            }
-            "success" -> {
-                val view = inflater.inflate(R.layout.layout_text_view_popup, null)
-                val textView: TextView = view.findViewById(R.id.text)
-                textView.text = textViewSelected
-                textView.setTextColor(Color.parseColor("#A81B1B"))
-                builder.setView(view)
-                    .setTitle("Success! Recognized Speech: \n")
-                    .setNegativeButton("done") { dialogInterface, i -> }
-//                    .setPositiveButton("ok") { dialogInterface, i ->
-//                        listener!!.deleteKeyPhrase()
-//                    }
-            }
-            "unrecognized" -> {
-                val view = inflater.inflate(R.layout.layout_text_view_popup, null)
-                val textView: TextView = view.findViewById(R.id.text)
-                textView.text = "Try again, or try something else"
-                builder.setView(view)
-                    .setTitle("Could not recognize speech with that key phrase \n")
-                    .setNegativeButton("Done") { dialogInterface, i -> }
-//                    .setPositiveButton("ok") { dialogInterface, i ->
-//                        listener!!.deleteKeyPhrase()
-            }
-//            }
-        }
-        return builder.create()
-    }
-
-//    override fun onAttach(context: Context) {
-//        super.onAttach(context)
-//        listener = try {
-//            context as Listener
-//        } catch (e: ClassCastException) {
-//            throw ClassCastException(
-//                context.toString() +
-//                    "must implement ExampleDialogListener"
-//            )
-//        }
-//    }
-}
