+++ app/src/main/java/com/metrostateics499/vre_app/view/CustomTextActivity.kt
@@ -22,18 +22,15 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
     private var textViewSelected: String = ""
     private var textViewSelectedBoolean: Boolean = false
     private var viewSelected: View? = null
-    private lateinit var customTextObjectSelected: CustomTextMessage
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        supportActionBar?.hide()
         setContentView(R.layout.activity_custom_text_message_menu)
 
         buttonAdd = findViewById<View>(R.id.buttonAdd) as Button
-//        buttonAdd!!.setOnClickListener { openPopUp(textViewSelected, "add") }
         buttonEdit = findViewById<View>(R.id.buttonEdit) as Button
-//        buttonEdit!!.setOnClickListener { checkSelectForEdit() }
         buttonDelete = findViewById<View>(R.id.buttonDelete) as Button
-//        buttonDelete!!.setOnClickListener { checkSelectForDelete() }
         buttonViewEntireText = findViewById<View>(R.id.buttonViewEntireText) as Button
         refreshList()
     }
@@ -64,34 +61,6 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
         buttonViewEntireText!!.setOnClickListener {
             openPopUp("null", "viewEntireText")
         }
-
-//        textViewSelectedBoolean = false
-//        textViewSelected = ""
-//
-//        val listview: ListView = findViewById(R.id.recyclerViewCustomText)
-//        val arrayAdapter = ArrayAdapter(
-//            this,
-//            android.R.layout.simple_list_item_1,
-//            Passing.customTextMessageList
-//        )
-//        listview.adapter = arrayAdapter
-//        listview.setOnItemClickListener { parent, view, position, _ ->
-//            viewSelected?.setBackgroundResource(
-//                androidx.appcompat.R.drawable
-//                    .abc_item_background_holo_light
-//            )
-//            Toast.makeText(
-//                this@CustomTextActivity,
-//                "You have selected " +
-//                    parent.getItemAtPosition(position),
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            textViewSelected = parent.getItemAtPosition(position).toString()
-//            customTextObjectSelected = parent.getItemAtPosition(position) as CustomTextMessage
-//            viewSelected = view
-//            textViewSelectedBoolean = true
-//            view.setBackgroundResource(androidx.appcompat.R.drawable.abc_list_pressed_holo_dark)
-//        }
     }
 
     private fun checkSelectForDelete(textViewSelectedBoolean: Boolean, textViewSelected: String) {
@@ -198,14 +167,5 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             }
             refreshList()
         }
-    }
-
-    private fun checkUniqueness(customTextString: String): Boolean {
-        for (item in Passing.customTextMessageList) {
-            if (item.textMessage == customTextString) {
-                return false
-            }
-        }
-        return true
     }
 }
