+++ app/src/main/java/com/metrostateics499/vre_app/view/EditEmergencyMessageActivity.kt
@@ -19,15 +19,13 @@ import kotlinx.android.synthetic.main.activity_edit_emergency_message.*
 import kotlinx.android.synthetic.main.activity_edit_emergency_message.view.*
 
 class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePopUps.Listener {
-    private lateinit var locationManager: LocationManager
-    private lateinit var tvGpsLocation: TextView
     private val locationPermissionCode = 2
 
     private var textViewSelected: String = ""
-//    private var emergencyMessageSelected = Passing.selectedEmergencyMessageSetup
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        supportActionBar?.hide()
         setContentView(R.layout.activity_edit_emergency_message)
 
         refreshRelativeLayout()
@@ -50,24 +48,14 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
         }
 
         relativeLayout2.setOnClickListener {
-//            textViewSelected = Passing.selectedEmergencyMessageSetup.keyPhrase.phrase
-//            openPopUp("keyphrase")
             goToKeyPhraseMenu()
         }
 
         relativeLayout3.setOnClickListener {
-//            textViewSelected = Passing.selectedEmergencyMessageSetup.keyPhrase.phrase
-//            openPopUp("keyphrase")
             goToCustomTextMenu()
         }
 
-//        relativeLayout3.setOnClickListener {
-//            textViewSelected = Passing.selectedEmergencyMessageSetup.customTextMessage.toString()
-//            openPopUp("customTextMessage")
-//        }
-
         relativeLayout4.setOnClickListener {
-//            textViewSelected = Passing.selectedEmergencyMessageSetup.customTextMessage.toString()
             goToContactsMenu()
         }
 
@@ -151,18 +139,6 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
         }
     }
 
-//    private fun getLocation() {
-//        locationManager = getSystemService(Context.LOCATION_SERVICE) as LocationManager
-//        if ((ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED)) {
-//            ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.ACCESS_FINE_LOCATION), locationPermissionCode)
-//        }
-//        locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 5000, 5f, this)
-//    }
-//    override fun onLocationChanged(location: Location) {
-//        tvGpsLocation = findViewById(R.id.textView)
-//        tvGpsLocation.text = "Latitude: " + location.latitude + " , Longitude: " + location.longitude
-//    }
-
     override fun onRequestPermissionsResult(
         requestCode: Int,
         permissions: Array<out String>,
@@ -213,24 +189,17 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
     }
     private fun refreshRelativeLayout2() {
         if (Passing.selectedEmergencyMessageSetup.selectedKeyPhraseList.isNotEmpty()) {
+            text_view_keyphrase_required.text = ""
             val textView2: TextView = findViewById(R.id.text_view_keyphrase)
             textView2.text =
                 (
                     Passing.selectedEmergencyMessageSetup.getKeyPhraseListString()
                     )
         } else if (Passing.selectedEmergencyMessageSetup.selectedKeyPhraseList.isEmpty()) {
+            text_view_keyphrase_required.text = "*"
             val textView2: TextView = findViewById(R.id.text_view_keyphrase)
             textView2.text = "Choose or Create Key Phrase(s)"
         }
-
-//        val textView2: TextView = findViewById(R.id.text_view_keyphrase)
-//        if (Passing.selectedEmergencyMessageSetup.keyPhrase.phrase.isNotEmpty()) {
-//            textView2.text = (
-//                (
-//                    Passing.selectedEmergencyMessageSetup.keyPhrase.phrase
-//                    )
-//                )
-//        }
     }
 
     private fun refreshRelativeLayout3() {
@@ -250,12 +219,14 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
 
     private fun refreshRelativeLayout4() {
         if (Passing.selectedEmergencyMessageSetup.selectedContactList.isNotEmpty()) {
+            text_view_contact_required.text = ""
             val textView4: TextView = findViewById(R.id.text_contact_list)
             textView4.text =
                 (
                     Passing.selectedEmergencyMessageSetup.getContactListNames()
                     )
         } else if (Passing.selectedEmergencyMessageSetup.selectedContactList.isEmpty()) {
+            text_view_contact_required.text = "*"
             val textView4: TextView = findViewById(R.id.text_contact_list)
             textView4.text = "Choose or Create Contacts"
         }
@@ -320,71 +291,6 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
         }
     }
 
-//    override fun editEmergencyMessageSetupKeyPhrase(inputPhrase: String) {
-//        if (inputPhrase.isEmpty()) {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity,
-//                "Key phrase can't be empty",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("keyphrase")
-//        } else if (inputPhrase == textViewSelected) {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity, "Make a change or click cancel",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("keyphrase")
-//        } else if (inputPhrase.isNotEmpty() &&
-//            checkKeyPhraseUniqueness(inputPhrase)
-//        ) {
-//            Passing.selectedEmergencyMessageSetup.keyPhrase.phrase = inputPhrase
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity, "Successfully Edited",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            refreshRelativeLayout2()
-//        } else {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity,
-//                "That Key Phrase already exists. " +
-//                    "Try something else or click cancel.",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("keyphrase")
-//        }
-//    }
-
-//    override fun editEmergencyMessageSetupCustomTextMessage(inputText: String) {
-//        if (inputText.isEmpty()) {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity,
-//                "Custom text message can't be empty",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("customTextMessage")
-//        } else if (inputText == textViewSelected) {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity, "Make a change or click cancel",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("customTextMessage")
-//        } else if (inputText.isNotEmpty()) {
-//            Passing.selectedEmergencyMessageSetup.customTextMessage.textMessage = inputText
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity, "Successfully Edited",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            refreshRelativeLayout3()
-//        } else {
-//            Toast.makeText(
-//                this@EditEmergencyMessageActivity,
-//                "That custom text message already exists. " +
-//                    "Try something else or click cancel.",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//            openPopUp("customTextMessage")
-//        }
-//    }
     override fun onPostResume() {
         super.onPostResume()
         refreshRelativeLayout2()
@@ -401,13 +307,4 @@ class EditEmergencyMessageActivity : AppCompatActivity(), EditEmergencyMessagePo
         }
         return true
     }
-
-//    private fun checkKeyPhraseUniqueness(phrase: String): Boolean {
-//        for (item in Passing.emergencyMessageSetupList) {
-//            if (item.keyPhrase.phrase.equals(phrase, true)) {
-//                return false
-//            }
-//        }
-//        return true
-//    }
 }
