+++ app/src/main/java/com/metrostateics499/vre_app/view/ListenSpeechActivity.kt
@@ -41,6 +41,7 @@ class ListenSpeechActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        supportActionBar?.hide()
         setContentView(R.layout.activity_listen_speech)
         initializeComponents()
         setListeners()
